% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/calc_spline.R
\name{calc.spline}
\alias{calc.spline}
\title{Calculate Spline Terms Based on rcs()}
\usage{
calc.spline(x, k1, k2, k3)
}
\arguments{
\item{x}{Numeric vector}

\item{k1}{Value of knot 1}

\item{k2}{Value of knot 2}

\item{k3}{Value of knot 3}
}
\value{
Numeric vector of X' terms.
}
\description{
Calculate X' values for restricted cubic spline term, eg, linear predictor = B0 + B1*X + B2*X'.
Based on Frank Harrell's Regression Modeling Strategies.
}
\examples{
calc.spline(rnorm(n = 100), k1 = -1.27, k2 = 0.02, k3 = 1.04)
}
\seealso{
\code{\link[rms]{rcs}} which this function assumes you use,
\code{\link[Hmisc]{rcspline.eval}} for getting knot locations
}

