% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/create_sampdata.R
\name{create.sampdata}
\alias{create.sampdata}
\title{Create Cluster Bootstrapped Data Sets}
\usage{
create.sampdata(org.data, id.var, n.sets, first.seed = 56)
}
\arguments{
\item{org.data}{Original data set, of class \code{data.frame}.}

\item{id.var}{Character string; name of subject identifier variable to sample.}

\item{n.sets}{Integer; number of final data sets desired.}

\item{first.seed}{Numeric (defaults to 56); set for reproducibility.}
}
\value{
List of \code{n.sets} data frames.
}
\description{
Using an original data frame with N subjects, creates data sets with all records from each of
N subject IDs sampled with replacement. Final data sets are saved in a list.
}
\examples{
df <- data.frame(id = sample(1:20, size = 100, replace = TRUE),
                 x1 = rnorm(n = 100),
                 x2 = rbinom(p = 0.75, n = 100, size = 1),
                 y = sample(LETTERS[1:3], size = 100, replace = TRUE))
df <- df[order(df$id),]
df$time <- unlist(lapply(1:length(unique(df$id)),
                         FUN = function(idnum){ 1:nrow(df[df$id == unique(df$id)[idnum],]) }))

## Using create.sampdata(), generate list of cluster bootstrapped data sets
bootdata.list <- create.sampdata(org.data = df,
                                 id.var = 'id',
                                 n.sets = 25)
}

